// src/types/response.ts
export interface RReviews {
  page: number;
  results: Review[];
  total_pages: number;
  total_results: number;
}
export interface RExternalId {
  imdb_id: string;
  facebook_id: string;
  instagram_id: string;
  twitter_id: string;
}
export interface RLists {
  page: number;
  results: List[];
  total_pages: number;
  total_results: number;
}

export interface RRecommendations {
  page: number;
  results: Movie[];
  total_pages: number;
  total_results: number;
}

export interface RKeywords {
  id?: number;
  keywords: { id: number; name: string }[];
}
export interface RTitles {
  titles: {
    iso_3166_1: string;
    title: string;
    type: string;
  }[];
}
export interface RVideos {
  results: {
    id: string;
    iso_639_1: string;
    iso_3166_1: string;
    key: string;
    name: string;
    site: string;
    size: number;
    type: string;
  }[];
}
// ...
export interface RImages {
  id?: number;
  backdrops: Img[];
  logos?: Img[];
  posters: Img[];
}
export interface Img {
  aspect_ratio: number;
  height: number;
  iso_639_1: string | null;
  file_path: string;
  vote_average: number;
  vote_count: number;
  width: number;
}
export interface RRelaseDates {
  results: {
    iso_3166_1: string;
    release_dates: [
      {
        certification: string;
        iso_639_1: string;
        note: string;
        release_date: string;
        type: number;
      },
    ];
  }[];
}
export interface RSimilar {
  page: number;
  results: MovieAndTv[];
  total_pages: number;
  total_results: number;
}
export type MovieAndTv = {
  backdrop_path: null | string;

  genre_ids: [number];

  id: number;
  name?: string;
  original_language: string;
  overview: string;
  popularity: number;
  poster_path: string;
  vote_average: number;
  vote_count: number;
  title?: string;
  original_title?: string;
  release_date?: string;
  adult: boolean;
  video?: boolean | string;
  origin_country?: [string];
  original_name?: string;
  first_air_date?: string;
};
// ...
type Review = {
  author: string;
  author_details: {
    name: string;
    username: string;
    avatar_path: string;
    rating: number;
  };
  content: string;
  created_at: string;
  id: string;
  updated_at: string;
  url: string;
};